change module structure: move validor from view to core
