+++ hw_lesson 19/flask_app/app1/core/Exceptions_register.py
