+++ hw_lesson 19/flask_app/app1/templates/app1/quote.html
@@ -12,7 +12,7 @@
     <div class="container">
     <h1>Quotes by Kanye West</h1>
     </div>
-    
+
     <ul>
     {% for i in req_list %}
         <option value>{{ i }}</option>
